rename: bank에서 cardCompany로 이름 변경
